requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
numpy==2.0.2
openai==1.61.1
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.15
contextual-client==0.5.1